-condense hex check -update errors
